+++ packages/learn-card-init/src/initializers/networkLearnCardFromSeed.ts
@@ -6,9 +6,7 @@ import { getDidKeyPlugin } from '@learncard/didkey-plugin';
 import { getEncryptionPlugin } from '@learncard/encryption-plugin';
 import { getVCPlugin } from '@learncard/vc-plugin';
 import { getVCTemplatesPlugin } from '@learncard/vc-templates-plugin';
-import { getCeramicPlugin } from '@learncard/ceramic-plugin';
 import { getLearnCloudPlugin } from '@learncard/learn-cloud-plugin';
-import { getIDXPlugin } from '@learncard/idx-plugin';
 import { expirationPlugin } from '@learncard/expiration-plugin';
 import { getEthereumPlugin } from '@learncard/ethereum-plugin';
 import { getVpqrPlugin } from '@learncard/vpqr-plugin';
@@ -17,7 +15,7 @@ import { getVerifyBoostPlugin, getLearnCardNetworkPlugin } from '@learncard/netw
 import { getLearnCardPlugin } from '@learncard/learn-card-plugin';
 
 import { NetworkLearnCardFromSeed } from '../types/LearnCard';
-import { defaultCeramicIDXArgs, defaultEthereumArgs } from '../defaults';
+import { defaultEthereumArgs } from '../defaults';
 
 /**
  * Generates a Network LearnCard Wallet from a 64 character seed string
@@ -35,7 +33,6 @@ export const networkLearnCardFromSeed = async ({
         unencryptedCustomFields = [],
         automaticallyAssociateDids = true,
     } = {},
-    ceramicIdx = defaultCeramicIDXArgs,
     didkit,
     allowRemoteContexts = false,
     ethereumConfig = defaultEthereumArgs,
@@ -59,11 +56,9 @@ export const networkLearnCardFromSeed = async ({
 
     const templateLc = await vcLc.addPlugin(getVCTemplatesPlugin());
 
-    const ceramicLc = await templateLc.addPlugin(await getCeramicPlugin(templateLc, ceramicIdx));
-
-    const cloudLc = await ceramicLc.addPlugin(
+    const cloudLc = await templateLc.addPlugin(
         await getLearnCloudPlugin(
-            ceramicLc,
+            templateLc,
             url,
             unencryptedFields,
             unencryptedCustomFields,
@@ -71,9 +66,7 @@ export const networkLearnCardFromSeed = async ({
         )
     );
 
-    const idxLc = await cloudLc.addPlugin(await getIDXPlugin(cloudLc, ceramicIdx));
-
-    const expirationLc = await idxLc.addPlugin(expirationPlugin(idxLc));
+    const expirationLc = await cloudLc.addPlugin(expirationPlugin(cloudLc));
 
     const ethLc = await expirationLc.addPlugin(getEthereumPlugin(expirationLc, ethereumConfig));