+++ services/learn-card-network/brain-service/src/helpers/learnCard.helpers.ts
@@ -4,9 +4,9 @@ import { generateLearnCard, LearnCard } from '@learncard/core';
 import { CryptoPlugin, CryptoPluginType } from '@learncard/crypto-plugin';
 import { DIDKitPlugin, DidMethod, getDidKitPlugin } from '@learncard/didkit-plugin';
 import { DidKeyPlugin, getDidKeyPlugin } from '@learncard/didkey-plugin';
+import { EncryptionPluginType, getEncryptionPlugin } from '@learncard/encryption-plugin';
 import { VCPlugin, getVCPlugin } from '@learncard/vc-plugin';
 import { VCTemplatePlugin, getVCTemplatesPlugin } from '@learncard/vc-templates-plugin';
-import { CeramicPlugin, getCeramicPlugin } from '@learncard/ceramic-plugin';
 import { ExpirationPlugin, expirationPlugin } from '@learncard/expiration-plugin';
 import { LearnCardPlugin, getLearnCardPlugin } from '@learncard/learn-card-plugin';
 import { getDidWebPlugin, DidWebPlugin } from '@learncard/did-web-plugin';
@@ -15,7 +15,9 @@ import { getDidWebPlugin, DidWebPlugin } from '@learncard/did-web-plugin';
 let didkitPromise: Promise<Buffer> | null = null;
 const getDidkitWasm = () => {
     if (!didkitPromise) {
-        didkitPromise = readFile(require.resolve('@learncard/didkit-plugin/dist/didkit_wasm_bg.wasm'));
+        didkitPromise = readFile(
+            require.resolve('@learncard/didkit-plugin/dist/didkit_wasm_bg.wasm')
+        );
     }
     return didkitPromise;
 };
@@ -29,9 +31,9 @@ export type SeedLearnCard = LearnCard<
         CryptoPluginType,
         DIDKitPlugin,
         DidKeyPlugin<DidMethod>,
+        EncryptionPluginType,
         VCPlugin,
         VCTemplatePlugin,
-        CeramicPlugin,
         ExpirationPlugin,
         LearnCardPlugin
     ]
@@ -42,9 +44,9 @@ export type DidWebLearnCard = LearnCard<
         CryptoPluginType,
         DIDKitPlugin,
         DidKeyPlugin<DidMethod>,
+        EncryptionPluginType,
         VCPlugin,
         VCTemplatePlugin,
-        CeramicPlugin,
         ExpirationPlugin,
         LearnCardPlugin,
         DidWebPlugin
@@ -86,13 +88,13 @@ export const getLearnCard = async (seed = process.env.SEED): Promise<SeedLearnCa
             await getDidKeyPlugin<DidMethod>(didkitLc, seed, 'key')
         );
 
-        const didkeyAndVCLc = await didkeyLc.addPlugin(getVCPlugin(didkeyLc));
+        const encryptionLc = await didkeyLc.addPlugin(await getEncryptionPlugin(didkeyLc));
 
-        const templateLc = await didkeyAndVCLc.addPlugin(getVCTemplatesPlugin());
+        const vcLc = await encryptionLc.addPlugin(getVCPlugin(encryptionLc));
 
-        const ceramicLc = await templateLc.addPlugin(await getCeramicPlugin(templateLc, {} as any));
+        const templateLc = await vcLc.addPlugin(getVCTemplatesPlugin());
 
-        const expirationLc = await ceramicLc.addPlugin(expirationPlugin(ceramicLc));
+        const expirationLc = await templateLc.addPlugin(expirationPlugin(templateLc));
 
         learnCards[seed] = await expirationLc.addPlugin(getLearnCardPlugin(expirationLc));
     }
@@ -122,13 +124,13 @@ export const getDidWebLearnCard = async (): Promise<DidWebLearnCard> => {
             await getDidKeyPlugin<DidMethod>(didkitLc, seed, 'key')
         );
 
-        const didkeyAndVCLc = await didkeyLc.addPlugin(getVCPlugin(didkeyLc));
+        const encryptionLc = await didkeyLc.addPlugin(await getEncryptionPlugin(didkeyLc));
 
-        const templateLc = await didkeyAndVCLc.addPlugin(getVCTemplatesPlugin());
+        const vcLc = await encryptionLc.addPlugin(getVCPlugin(encryptionLc));
 
-        const ceramicLc = await templateLc.addPlugin(await getCeramicPlugin(templateLc, {} as any));
+        const templateLc = await vcLc.addPlugin(getVCTemplatesPlugin());
 
-        const expirationLc = await ceramicLc.addPlugin(expirationPlugin(ceramicLc));
+        const expirationLc = await templateLc.addPlugin(expirationPlugin(templateLc));
 
         const lcLc = await expirationLc.addPlugin(getLearnCardPlugin(expirationLc));