+++ packages/plugins/chapi/CHANGELOG.md
@@ -1,5 +1,13 @@
 # learn-card-core
 
+## 1.0.48
+
+### Patch Changes
+
+-   Updated dependencies [[`a945a7086a0848b04f9f2c8b1588e9e3826edeff`](https://github.com/learningeconomy/LearnCard/commit/a945a7086a0848b04f9f2c8b1588e9e3826edeff)]:
+    -   @learncard/core@9.3.20
+    -   @learncard/didkit-plugin@1.5.6
+
 ## 1.0.47
 
 ### Patch Changes