+++ services/meta-mask-snap/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @learncard/meta-mask-snap
 
+## 1.0.71
+
+### Patch Changes
+
+-   Updated dependencies [[`a945a7086a0848b04f9f2c8b1588e9e3826edeff`](https://github.com/learningeconomy/LearnCard/commit/a945a7086a0848b04f9f2c8b1588e9e3826edeff)]:
+    -   @learncard/core@9.3.20
+
 ## 1.0.70
 
 ### Patch Changes